+++ perseo-quality/docs/source/changelog.rst
@@ -9,6 +9,7 @@ v1.0.2
 - Fixed graphs name and title for Point Target Analysis and Radiometric Analysis Block-Wise
 - Added ``sensor_name`` property to quality input channel protocol
 - Added ``sensor``, ``product`` and ``acquisition_mode`` info to Point Target Analysis output .csv
+- Removed ``pulse_latch_time`` and ``swst_changes`` from protocol definition
 
 v1.0.1
 ------